refactor: rename resource files.
