+++ test/tokens.spec.js
